WIP, additional refactoring. Expose more of the datastructure internals, and split up modules to make them more digestible.
