+++ crucible/src/Lang/Crucible/Simulator/ExecutionTree.hs
@@ -8,9 +8,13 @@
 -- Stability        : provisional
 --
 -- Execution trees record the state of the simulator as it explores
--- execution paths through a program.
+-- execution paths through a program.  This module defines the
+-- collection of datatypes that record the state of a running simulator
+-- and basic lenses and accessors for these types. See
+-- "Lang.Crucible.Simulator.Operations" for the definitions of operations
+-- that manipulate these datastructures to drive them through the simulator
+-- state machine.
 ------------------------------------------------------------------------
-
 {-# LANGUAGE ConstraintKinds #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DeriveDataTypeable #-}
@@ -33,100 +37,112 @@ module Lang.Crucible.Simulator.ExecutionTree
     GlobalPair(..)
   , gpValue
   , gpGlobals
+
+    -- * TopFrame
   , TopFrame
   , crucibleTopFrame
+
+    -- * CrucibleBranchTarget
+  , CrucibleBranchTarget(..)
+  , ppBranchTarget
+
     -- * AbortedResult
   , AbortedResult(..)
-    -- * PartialResult
-  , PartialResult(..)
-    -- * Frames
   , SomeFrame(..)
   , filterCrucibleFrames
   , arFrames
   , ppExceptionContext
-  , abortTree
-  , defaultErrorHandler
-    -- * ExecResult
+
+    -- * Partial result
+  , PartialResult(..)
+  , partialValue
+
+    -- * Execution results
   , ExecResult(..)
+
+    -- * Execution intermediate states
   , ExecState(..)
   , ExecCont
 
-    -- * ActiveTree
-  , ActiveTree(ActiveTree)
-  , singletonTree
-  , actContext
-  , activeFrames
-    -- ** Active tree helpers.
-  , actFrame
-    -- ** ValueFromFrame
-  , ReturnHandler
+    -- * Simulator context trees
+    -- ** Main context data structures
+  , ValueFromValue(..)
+  , ValueFromFrame(..)
+  , PendingPartialMerges(..)
+
+    -- ** Paused Frames
+  , PausedFrame(..)
+  , pausedFrame
+  , resume
+
+    -- ** Sibling paths
+  , VFFOtherPath(..)
+  , FrameRetType
+
+    -- ** ReturnHandler
+  , ReturnHandler(..)
   , returnToOverride
   , returnToCrucible
   , tailReturnToCrucible
-  , continue
-  , jumpToBlock
-  , runOverride
 
-  , returnAndMerge
-  , returnValue
-  , FrameRetType
+    -- * ControlResumption
+  , ControlResumption(..)
+  , ResolvedJump(..)
 
-  , isSingleCont
-  , pathConditions
-    -- * Branch information
-  , PausedValue(..)
-  , PausedFrame(..)
+    -- * ActiveTree
+  , ActiveTree(..)
+  , singletonTree
+  , activeFrames
+  , actContext
+  , actFrame
 
-    -- ** Branch and merge at return
-  , intra_branch
-  , SomeLabel(..)
-    -- * High level operations.
-  , replaceTailFrame
-  , getIntraFrameBranchTarget
-  , performIntraFrameMerge
-  , callFn
-  , abortExec
   , extractCurrentPath
-  , branchConditions
-    -- * SimState
-  , SimState(..)
-  , CrucibleState
-  , stateTree
-  , IsSymInterfaceProof
-  , ErrorHandler(..)
-  , CrucibleBranchTarget(..)
-  , errorHandler
-  , stateContext
-  , stateCrucibleFrame
-    -- * SimContext
+  , asContFrame
+
+--  , branchConditions
+--  , pathConditions
+
+    -- * Simulator context
+    -- ** Function bindings
   , Override(..)
+  , FnState(..)
+  , FunctionBindings
+
+    -- ** Extensions
+  , ExtensionImpl(..)
+  , EvalStmtFunc
+
+    -- ** SimContext record
+  , IsSymInterfaceProof
   , SimContext(..)
   , initSimContext
   , ctxSymInterface
   , functionBindings
   , cruciblePersonality
-  , FunctionBindings
-  , FnState(..)
-  , ExtensionImpl(..)
-  , EvalStmtFunc
-    -- * Utilities
+
+    -- * SimState
+  , AbortHandler(..)
+  , SimState(..)
+  , CrucibleState
+  , initSimState
+
+    -- ** Lenses and accessors
+  , stateTree
+  , abortHandler
+  , stateContext
+  , stateCrucibleFrame
   , stateSymInterface
   , stateSolverProof
   , stateIntrinsicTypes
   , stateOverrideFrame
   , stateGetConfiguration
-  , runErrorHandler
-  , runGenericErrorHandler
   ) where
 
 import           Control.Lens
 import           Control.Monad.Reader
 import           Control.Monad.ST (RealWorld)
-import           Data.Monoid ((<>))
 import qualified Data.Parameterized.Context as Ctx
-import           Data.Parameterized.Some
 import           Data.Sequence (Seq)
-import qualified Data.Sequence as Seq
 import           Data.Type.Equality hiding (sym)
 import           System.Exit (ExitCode)
 import           System.IO
@@ -134,14 +150,13 @@ import qualified Text.PrettyPrint.ANSI.Leijen as PP
 
 import           What4.Config
 import           What4.Interface
-import           What4.FunctionName (FunctionName)
+import           What4.FunctionName (FunctionName, startFunctionName)
 import           What4.ProgramLoc
 
 import           Lang.Crucible.Backend
 import           Lang.Crucible.CFG.Core
 import           Lang.Crucible.CFG.Extension
 import           Lang.Crucible.FunctionHandle (FnHandleMap, HandleAllocator)
-import           Lang.Crucible.Panic(panic)
 import           Lang.Crucible.Simulator.CallFrame
 import           Lang.Crucible.Simulator.Evaluation
 import           Lang.Crucible.Simulator.Frame
@@ -150,23 +165,46 @@ import           Lang.Crucible.Simulator.Intrinsics
 import           Lang.Crucible.Simulator.RegMap
 
 
+------------------------------------------------------------------------
+-- GlobalPair
+
+-- | A value of some type 'v' together with a global state.
+data GlobalPair sym (v :: *) =
+   GlobalPair { _gpValue :: !v
+              , _gpGlobals :: !(SymGlobalState sym)
+              }
+
+-- | The value stored in the global pair.
+gpValue :: Lens (GlobalPair sym u) (GlobalPair sym v) u v
+gpValue = lens _gpValue (\s v -> s { _gpValue = v })
+
+-- | The globals stored in the global pair.
+gpGlobals :: Simple Lens (GlobalPair sym u) (SymGlobalState sym)
+gpGlobals = lens _gpGlobals (\s v -> s { _gpGlobals = v })
 
 
--- | @swap_unless b (x,y)@ returns @(x,y)@ when @b@ is @True@ and
--- @(y,x)@ when @b@ if @False@.
-swap_unless :: Bool -> (a, a) -> (a,a)
-swap_unless True p = p
-swap_unless False (x,y) = (y,x)
-{-# INLINE swap_unless #-}
+------------------------------------------------------------------------
+-- TopFrame
 
--- | Return assertion where predicate equals a constant
-predEqConst :: IsExprBuilder sym => sym -> Pred sym -> Bool -> IO (Pred sym)
-predEqConst _   p True  = return p
-predEqConst sym p False = notPred sym p
+-- | The currently-exeucting frame plus the global state associated with it.
+type TopFrame sym ext f a = GlobalPair sym (SimFrame sym ext f a)
+
+-- | A lens to access the Crucible call frame inside a 'TopFrame'
+crucibleTopFrame ::
+  Lens (TopFrame sym ext (CrucibleLang blocks r) ('Just args))
+       (TopFrame sym ext (CrucibleLang blocks r) ('Just args'))
+       (CallFrame sym ext blocks r args)
+       (CallFrame sym ext blocks r args')
+crucibleTopFrame = gpValue . crucibleSimFrame
+{-# INLINE crucibleTopFrame #-}
+
 
 ------------------------------------------------------------------------
 -- CrucibleBranchTarget
 
+-- | A 'CrucibleBranchTarget' identifies a program location that is a
+--   potential join point.  Each label is a merge point, and there is
+--   an additional implicit join point at function returns.
 data CrucibleBranchTarget blocks (args :: Maybe (Ctx CrucibleType)) where
    BlockTarget  :: !(BlockID blocks args)
                 -> CrucibleBranchTarget blocks ('Just args)
@@ -186,83 +224,44 @@ ppBranchTarget ReturnTarget = "return"
 
 
 ------------------------------------------------------------------------
--- GlobalPair
+-- AbortedResult
 
--- | A value with a global state associated.
-data GlobalPair sym (v :: *) =
-   GlobalPair { _gpValue :: !v
-              , _gpGlobals :: !(SymGlobalState sym)
-              }
+-- | An execution path that was prematurely aborted.  Note, an abort
+--   does not necessarily inidicate an error condition.  An execution
+--   path might abort because it became infeasible (inconsistent path
+--   conditions), because the program called an exit primitive, or
+--   because of a true error condition (e.g., a failed assertion).
+data AbortedResult sym ext where
+  -- | A single aborted execution with the execution state at time of the abort and the reason.
+  AbortedExec ::
+    !AbortExecReason ->
+    !(GlobalPair sym (SimFrame sym ext l args)) ->
+    AbortedResult sym ext
 
--- | The value stored in the global pair.
-gpValue :: Lens (GlobalPair sym u) (GlobalPair sym v) u v
-gpValue = lens _gpValue (\s v -> s { _gpValue = v })
+  -- | An aborted execution that was ended by a call to 'exit'.
+  AbortedExit ::
+    !ExitCode ->
+    AbortedResult sym ext
 
--- | The globals stored in the global pair.
-gpGlobals :: Simple Lens (GlobalPair sym u) (SymGlobalState sym)
-gpGlobals = lens _gpGlobals (\s v -> s { _gpGlobals = v })
-
--- | Merge two globals together.
-mergeGlobalPair :: MuxFn p v
-                -> MuxFn p (SymGlobalState sym)
-                -> MuxFn p (GlobalPair sym v)
-mergeGlobalPair merge_fn global_fn c x y =
-  GlobalPair <$> merge_fn  c (x^.gpValue) (y^.gpValue)
-             <*> global_fn c (x^.gpGlobals) (y^.gpGlobals)
-
-------------------------------------------------------------------------
--- TopFrame
-
--- | A frame plus the global state associated with it.
-type TopFrame sym ext f a = GlobalPair sym (SimFrame sym ext f a)
-
-crucibleTopFrame ::
-  Lens (TopFrame sym ext (CrucibleLang blocks r) ('Just args))
-       (TopFrame sym ext (CrucibleLang blocks r) ('Just args'))
-       (CallFrame sym ext blocks r args)
-       (CallFrame sym ext blocks r args')
-crucibleTopFrame = gpValue . crucibleSimFrame
-
-stateCrucibleFrame ::
-  Lens (SimState p sym ext rtp (CrucibleLang blocks r) ('Just a))
-       (SimState p sym ext rtp (CrucibleLang blocks r) ('Just a'))
-       (CallFrame sym ext blocks r a)
-       (CallFrame sym ext blocks r a')
-stateCrucibleFrame = stateTree . actFrame . crucibleTopFrame
-{-# INLINE stateCrucibleFrame #-}
-
-
+  -- | Two separate threads of execution aborted after a symbolic branch,
+  --   possibly for different reasons
+  AbortedBranch ::
+    !(Pred sym)              {- The symbolic condition -} ->
+    !(AbortedResult sym ext) {- The abort that occurred along the 'true' branch -} ->
+    !(AbortedResult sym ext) {- The abort that occurred along the 'false' branch -} ->
+    AbortedResult sym ext
 
 ------------------------------------------------------------------------
 -- SomeFrame
 
--- | This represents an execution frame.
+-- | This represents an execution frame where its frame type
+--   and arguments have been hidden.
 data SomeFrame (f :: fk -> argk -> *) = forall l a . SomeFrame !(f l a)
 
-filterCrucibleFrames :: SomeFrame (SimFrame sym ext)
-                     -> Maybe ProgramLoc
+-- | Return the program locations of all the Crucible frames.
+filterCrucibleFrames :: SomeFrame (SimFrame sym ext) -> Maybe ProgramLoc
 filterCrucibleFrames (SomeFrame (MF f)) = Just (frameProgramLoc f)
 filterCrucibleFrames _ = Nothing
-
-------------------------------------------------------------------------
--- AbortedResult
-
--- | An execution that was prematurely aborted.
-data AbortedResult sym ext where
-  -- A single aborted execution with the execution state at time of error,
-  -- and the reason for the error.
-  AbortedExec :: !AbortExecReason
-              -> !(GlobalPair sym (SimFrame sym ext l args))
-              -> AbortedResult sym ext
-  -- An aborted execution that was ended by a call to 'exit'
-  AbortedExit :: !ExitCode
-              -> AbortedResult sym ext
-
-  -- An entire aborted branch.
-  AbortedBranch :: !(Pred sym)
-                -> !(AbortedResult sym ext)
-                -> !(AbortedResult sym ext)
-                -> AbortedResult sym ext
 
 -- | Iterate over frames in the result.
 arFrames :: Simple Traversal (AbortedResult sym ext) (SomeFrame (SimFrame sym ext))
@@ -274,7 +273,7 @@ arFrames h (AbortedBranch p r s) =
   AbortedBranch p <$> arFrames h r
                   <*> arFrames h s
 
--- | Print the exception
+-- | Print an exception context
 ppExceptionContext :: [SomeFrame (SimFrame sym ext)] -> PP.Doc
 ppExceptionContext [] = PP.empty
 ppExceptionContext frames = PP.vcat (map pp (init frames))
@@ -288,43 +287,27 @@ ppExceptionContext frames = PP.vcat (map pp (init frames))
    pp (SomeFrame (RF _v)) =
       PP.text ("While returning value")
 
--- | Abort the current execution.
-abortTree ::
-  IsSymInterface sym =>
-  AbortExecReason ->
-  ExecCont p sym ext rtp f args
-abortTree e = do
-  t   <- view stateTree
-  cfg <- asks stateGetConfiguration
-  ctx <- view stateContext
-  v <- liftIO (getOpt =<< getOptionSetting verbosity cfg)
-  when (v > 0) $ do
-    let frames = activeFrames t
-    let msg = ppAbortExecReason e PP.<$$>
-              PP.indent 2 (ppExceptionContext frames)
-    -- Print error message.
-    liftIO (hPrint (printHandle ctx) msg)
-  -- Switch to new frame.
-  abortExec e
-
-defaultErrorHandler :: IsSymInterface sym => ErrorHandler p sym ext rtp
-defaultErrorHandler = EH abortTree
 
 ------------------------------------------------------------------------
 -- PartialResult
 
--- | 'PartialResult' contains a value and global variables along with an
--- optional aborted result.
+-- | A 'PartialResult' represents the result of a computation that
+--   might be only partially defined.  If the result is a 'TotalResult',
+--   the the result is fully defined; however if it is a
+--   'PartialResult', then some of the computation paths that led to
+--   this result aborted for some reason, and the resulting value is
+--   only defined if the associated condition is true.
 data PartialResult sym ext (v :: *)
 
     -- | A 'TotalRes' indicates that the the global pair is always defined.
    = TotalRes !(GlobalPair sym v)
 
     {- | 'PartialRes' indicates that the global pair may be undefined
-        under some circusmstances.  The predicate specifies when the
-        'GlobalPair' is guaranteed to be defined.
+        under some circusmstances.  The predicate specifies under what
+        conditions the 'GlobalPair' is defined.
         The 'AbortedResult' describes the circumstances under which
-        the result would be partial. -}
+        the result would be partial.
+     -}
    | PartialRes !(Pred sym)               -- if true, global pair is defined
                 !(GlobalPair sym v)       -- the value
                 !(AbortedResult sym ext)  -- failure cases (when pred. is false)
@@ -340,201 +323,113 @@ partialValue f (TotalRes x) = TotalRes <$> f x
 partialValue f (PartialRes p x r) = (\y -> PartialRes p y r) <$> f x
 {-# INLINE partialValue #-}
 
-mergeAbortedBranch :: Pred sym
-                   -> AbortedResult sym ext
-                   -> AbortedResult sym ext
-                   -> AbortedResult sym ext
-mergeAbortedBranch _ (AbortedExit ec) _ = AbortedExit ec
-mergeAbortedBranch _ _ (AbortedExit ec) = AbortedExit ec
-mergeAbortedBranch c q r = AbortedBranch c q r
-
-mergePartialAndAbortedResult ::
-  IsExprBuilder sym =>
-  sym ->
-  Pred sym {- ^ This needs to hold to avoid the aborted result -} ->
-  PartialResult sym ext v ->
-  AbortedResult sym ext ->
-  IO (PartialResult sym ext v)
-mergePartialAndAbortedResult sym c ar r =
-  case ar of
-    TotalRes gp -> return $! PartialRes c gp r
-    PartialRes d gp q ->
-      do e <- andPred sym c d
-         return $! PartialRes e gp (mergeAbortedBranch c q r)
-
-mergePartialResult ::
-  IsSymInterface sym =>
-  SimState p sym ext f root args ->
-  CrucibleBranchTarget blocks args ->
-  MuxFn (Pred sym)
-     (PartialResult sym ext (SimFrame sym ext (CrucibleLang blocks ret) args))
-mergePartialResult s tgt pp x y =
-  let sym       = stateSymInterface s
-      iteFns    = stateIntrinsicTypes s
-      merge_val = mergeCrucibleFrame sym iteFns tgt
-      merge_fn  = mergeGlobalPair merge_val (globalMuxFn sym iteFns)
-  in
-  case x of
-    TotalRes cx ->
-      case y of
-        TotalRes cy ->
-          TotalRes <$> merge_fn pp cx cy
-
-        PartialRes py cy fy ->
-          PartialRes <$> orPred sym pp py
-                     <*> merge_fn pp cx cy
-                     <*> pure fy
-
-    PartialRes px cx fx ->
-      case y of
-        TotalRes cy ->
-          do pc <- notPred sym pp
-             PartialRes <$> orPred sym pc px
-                        <*> merge_fn pp cx cy
-                        <*> pure fx
-
-        PartialRes py cy fy ->
-          PartialRes <$> itePred sym pp px py
-                     <*> merge_fn pp cx cy
-                     <*> pure (AbortedBranch pp fx fy)
 
 ------------------------------------------------------------------------
 -- ExecResult
 
--- | Executions that have completed either due to (partial or total) successful execution
---   completion or by an error.
---
---   The included predicate indicates the path condition that resulted in the
---   given result.
+-- | Executions that have completed either due to (partial or total)
+--   successful completion or by some abort condition.
 data ExecResult p sym ext (r :: *)
-   = FinishedResult !(SimContext p sym ext) !(PartialResult sym ext r)
+   = -- | At least one exeuction path resulted in some return result.
+     FinishedResult !(SimContext p sym ext) !(PartialResult sym ext r)
+     -- | All execution paths resulted in an abort condition, and there is
+     --   no result to return.
    | AbortedResult  !(SimContext p sym ext) !(AbortedResult sym ext)
 
+-----------------------------------------------------------------------
+-- ExecState
+
+-- | An 'ExecState' represents an intermediate state of executing a
+--   Crucible program.  The Crucible simulator executes by transistioning
+--   between these different states until it results in a 'ResultState',
+--   indicating the program has completed.
 data ExecState p sym ext (rtp :: *)
-   =   ResultState
+   = -- | The 'ResultState' is used to indicate that the program has completed.
+       ResultState
          !(ExecResult p sym ext rtp)
 
-   | forall f a.
+   | -- | An abort state indicates that the included 'SimState' encountered
+     --   an abort event while executing its next step.  The state needs to
+     --   be unwound to its nearest enclosing branch point and resumed.
+     forall f a.
        AbortState
          !AbortExecReason
          !(SimState p sym ext rtp f a)
 
-   | forall blocks r args.
+   | -- | A running state indicates the included 'SimState' is ready to enter
+     --   and execute a Crucible basic block, or to resume a basic block
+     --   from a call site.
+     forall blocks r args.
        RunningState
          !(SimState p sym ext rtp (CrucibleLang blocks r) ('Just args))
 
-   | forall args ret.
+   | -- | An override state indicates the included 'SimState' is prepared to
+     --   execute a code override.
+     forall args ret.
        OverrideState
          !(Override p sym ext args ret)
          !(SimState p sym ext rtp (OverrideLang args ret) 'Nothing)
 
-   | forall blocks r args.
+   | -- | A control transfer state occurs when the included 'SimState' is
+     --   in the process of transfering control to the included 'CrucibleBranchTarget'.
+     --   During this process, paths may have to be merged.  If several branches
+     --   must merge at the same control point, this state may be entered several
+     --   times in succession before returning to a 'RunningState'.
+     forall blocks r args.
        ControlTransferState
-         !(ExecCont p sym ext rtp (CrucibleLang blocks r) args)
+         !(ControlResumption p sym ext rtp blocks r args)
+           {- Action to resume after all pending merges have been resolved -}
          !(CrucibleBranchTarget blocks args)
+           {- Target of the control-flow transfer -}
          !(SimState p sym ext rtp (CrucibleLang blocks r) args)
+           {- State of the simulator prior to the control-flow transfer -}
 
-------------------------------------------------------------------------
--- A Paused value and ExecCont
-
--- | An action which will construct an ExecState given a global state.
---   Such continuations correspond to a single step of execution.
+-- | An action which will construct an 'ExecState' given a current
+--   'SimState'. Such continuations correspond to a single transition
+--   of the simulator transition system.
 type ExecCont p sym ext r f a =
   ReaderT (SimState p sym ext r f a) IO (ExecState p sym ext r)
 
--- | This is essentially a closure denoting a type of continuation that needs a
--- global state to run, but currently has a value that it will use to generate
--- the state, along with a solver state used to configure the state of the
--- solver interface.
-data PausedValue p sym ext root f args (v :: *)
-   = PausedValue { _pausedValue :: !v
-                 , resume :: !(ExecCont p sym ext root f args)
-                   -- ^ Function to run.
-                 }
+data ResolvedJump sym blocks
+  = forall args.
+      ResolvedJump
+        !(Pred sym)
+        !(BlockID blocks args)
+        !(RegMap sym args)
 
-pausedValue :: Lens (PausedValue p sym ext root f a u)
-                    (PausedValue p sym ext root f a v)
-                    u
-                    v
-pausedValue = lens _pausedValue (\s v -> s { _pausedValue = v })
-
-resumed :: Simple Lens (PausedValue p sym ext root f a u)
-                       (ExecCont    p sym ext root f a)
-resumed = lens resume (\s v -> s { resume = v })
-
-------------------------------------------------------------------------
--- ReurnHandler and CallerCont
-
--- | This type names a simulator frame, paired with a continuation
--- to be used when returning from the function.
-
-data ReturnHandler top_return p sym ext r f args new_args where
-  ReturnToOverride ::
-    (ret -> ExecCont p sym ext rtp (OverrideLang args r) 'Nothing)
-      {- ^ Remaining override code to run when the return value becomse available -} ->
-    ReturnHandler ret p sym ext rtp (OverrideLang args r) 'Nothing 'Nothing
-
-  ReturnToCrucible ::
-    TypeRepr ret {- ^ Type of the return value -} ->
-    StmtSeq ext blocks r (ctx ::> ret) {- ^ Remaining statements to execute -} ->
-    ReturnHandler (RegEntry sym ret)
-      p sym ext root (CrucibleLang blocks r) ('Just ctx) ('Just (ctx ::> ret))
-
-  TailReturnToCrucible ::
-    ReturnHandler (RegEntry sym r)
-      p sym ext root (CrucibleLang blocks r) ctx 'Nothing
-
-
-returnToOverride ::
-  (ret -> SimState p sym ext rtp (OverrideLang args r) 'Nothing -> IO (ExecState p sym ext rtp))
-    {- ^ Remaining override code to run when the return value becomse available -} ->
-  ReturnHandler ret p sym ext rtp (OverrideLang args r) 'Nothing 'Nothing
-returnToOverride c = ReturnToOverride (ReaderT . c)
-
-returnToCrucible :: forall p sym ext root ret blocks r ctx.
-  TypeRepr ret {- ^ Type of the return value -} ->
-  StmtSeq ext blocks r (ctx ::> ret) {- ^ Remaining statements to execute -} ->
-  ReturnHandler (RegEntry sym ret)
-    p sym ext root (CrucibleLang blocks r) ('Just ctx) ('Just (ctx ::> ret))
-returnToCrucible = ReturnToCrucible
-
-tailReturnToCrucible :: forall p sym ext root blocks ctx r.
-  ReturnHandler (RegEntry sym r)
-    p sym ext root (CrucibleLang blocks r) ctx 'Nothing
-tailReturnToCrucible = TailReturnToCrucible
-
-runOverride ::
-  Override p sym ext args ret ->
-  ExecCont p sym ext rtp (OverrideLang args ret) 'Nothing
-runOverride o = ReaderT (return . OverrideState o)
-
-continue :: ExecCont p sym ext rtp (CrucibleLang blocks r) ('Just a)
-continue = ReaderT (return . RunningState)
-
-jumpToBlock :: (IsSymInterface sym, IsSyntaxExtension ext)
-             => BlockID blocks args
-             -> RegMap sym args
-             -> ExecCont p sym ext rtp (CrucibleLang blocks r) ('Just a)
-jumpToBlock block_id args =
-  withReaderT
-    (stateCrucibleFrame %~ setFrameBlock block_id args)
-    (checkForIntraFrameMerge continue (BlockTarget block_id))
-{-# INLINE jumpToBlock #-}
+data ControlResumption p sym ext rtp blocks r args where
+  ContinueResumption   :: ControlResumption p sym ext rtp blocks r ('Just args)
+  ReturnResumption     :: ControlResumption p sym ext rtp blocks r 'Nothing
+  CheckMergeResumption :: 
+    ControlResumption p sym ext root blocks r args ->
+    CrucibleBranchTarget blocks args ->
+    ControlResumption p sym ext root blocks r args
+  SwitchResumption ::
+    CrucibleBranchTarget blocks args' ->
+    [ResolvedJump sym blocks] ->
+    ControlResumption p sym ext root blocks r ('Just args)
 
 ------------------------------------------------------------------------
--- Recursive data types
+-- Paused Frame
 
-type PausedPartialFrame p sym ext root f args
-   = PausedValue p sym ext root f args (PartialResult sym ext (SimFrame sym ext f args))
+data PausedFrame p sym ext root b r args
+   = PausedFrame
+     { _pausedFrame  :: !(PartialResult sym ext (SimFrame sym ext (CrucibleLang b r) args))
+     , _resume       :: !(ControlResumption p sym ext root b r args)
+     }
 
--- | An active execution tree contains at least one active execution.
--- The data structure is organized so that the current execution
--- can be accessed rapidly.
-data ActiveTree p sym ext root (f :: *) args
-   = ActiveTree
-      { _actContext :: !(ValueFromFrame p sym ext root f)
-      , _actResult  :: !(PartialResult sym ext (SimFrame sym ext f args))
-      }
+pausedFrame ::
+  Simple Lens
+    (PausedFrame p sym ext root b r args)
+    (PartialResult sym ext (SimFrame sym ext (CrucibleLang b r) args))
+pausedFrame = lens _pausedFrame (\ppf v -> ppf{ _pausedFrame = v })
+
+resume ::
+  Simple Lens
+    (PausedFrame p sym ext root b r args)
+    (ControlResumption p sym ext root b r args)
+resume = lens _resume (\ppf r -> ppf{ _resume = r })    
+
 
 -- | This describes what to do in an inactive path.
 data VFFOtherPath p sym ext ret blocks r args
@@ -543,15 +438,18 @@ data VFFOtherPath p sym ext ret blocks r args
    = forall o_args.
       VFFActivePath
         !(Maybe ProgramLoc) {- Location of branch target -}
-        !(PausedPartialFrame p sym ext ret (CrucibleLang blocks r) o_args)
+        !(PausedFrame p sym ext ret blocks r o_args)
 
      -- | This is a completed execution path.
    | VFFCompletePath
         !(Seq (Assumption sym))
         {- Assumptions that we collected while analyzing the branch -}
+        !(PausedFrame p sym ext ret blocks r args)
 
-        !(PausedPartialFrame p sym ext ret (CrucibleLang blocks r) args)
 
+type family FrameRetType (f :: *) :: CrucibleType where
+  FrameRetType (CrucibleLang b r) = r
+  FrameRetType (OverrideLang b r) = r
 
 
 {- | This type contains information about the current state of the exploration
@@ -593,7 +491,7 @@ data ValueFromFrame p sym ext (ret :: *) (f :: *)
       !(VFFOtherPath p sym ext ret blocks r args)
       {- Info about the state of the other branch.
          If the other branch is "VFFActivePath", then we still
-         need to prcess it;  if it is "VFFCompletePath", then
+         need to process it;  if it is "VFFCompletePath", then
          it is finsihed, and so once we are done then we go back to the
          outer context. -}
 
@@ -631,7 +529,6 @@ data PendingPartialMerges =
   | NeedsToBeAborted
 
 
--- | value from value denotes
 data ValueFromValue p sym ext (ret :: *) (top_return :: *)
 
   -- VFVCall denotes a return to a given frame.
@@ -697,7 +594,8 @@ ppValueFromValue vfv =
       PP.pretty ctx
     VFVEnd -> PP.text "root"
 
-------------------------------------------------------------------------
+
+-----------------------------------------------------------------------
 -- parentFrames
 
 -- | Return parents frames in reverse order.
@@ -715,487 +613,80 @@ vfvParents c0 =
     VFVPartial c _ _ -> vfvParents c
     VFVEnd -> []
 
-------------------------------------------------------------------------
--- Utilities
 
-pushCrucibleFrame :: forall sym ext b r a
-                  .  IsSymInterface sym
-                  => sym
-                  -> IntrinsicTypes sym
-                  -> SimFrame sym ext (CrucibleLang b r) a
-                  -> IO (SimFrame sym ext (CrucibleLang b r) a)
-pushCrucibleFrame sym muxFns (MF x) = do
-  r' <- pushBranchRegs sym muxFns (x^.frameRegs)
-  return $! MF (x & frameRegs .~ r')
-pushCrucibleFrame sym muxFns (RF x) = do
-  x' <- pushBranchRegEntry sym muxFns x
-  return $! RF x'
+-- ------------------------------------------------------------------------
+-- -- pathConditions
 
+-- -- | Return list of conditions along current execution path.
+-- pathConditions :: ValueFromFrame p sym ext r a
+--                -> [Pred sym]
+-- pathConditions c0 =
+--   case c0 of
+--     VFFBranch   c _ _ p _ _ -> p : pathConditions c
+--     VFFPartial  c p _ _   -> p : pathConditions c
+--     VFFEnd vfv            -> vfvConditions vfv
 
-popCrucibleFrame :: IsSymInterface sym
-                 => sym
-                 -> IntrinsicTypes sym
-                 -> SimFrame sym ext (CrucibleLang b r') a'
-                 -> IO (SimFrame sym ext (CrucibleLang b r') a')
-popCrucibleFrame sym intrinsicFns (MF x') =
-  do r' <- abortBranchRegs sym intrinsicFns (x'^.frameRegs)
-     return $! MF (x' & frameRegs .~ r')
-
-popCrucibleFrame sym intrinsicFns (RF x') =
-  RF <$> abortBranchRegEntry sym intrinsicFns x'
-
-
-fromJustCallFrame :: SimFrame sym ext (CrucibleLang b r) ('Just a)
-                  -> CallFrame sym ext b r a
-fromJustCallFrame (MF x) = x
-
-fromNothingCallFrame :: SimFrame sym ext (CrucibleLang b r) 'Nothing
-                     -> RegEntry sym r
-fromNothingCallFrame (RF x) = x
-
-
-mergeCrucibleFrame :: IsSymInterface sym
-                   => sym
-                   -> IntrinsicTypes sym
-                   -> CrucibleBranchTarget blocks args -- ^ Target of branch
-                   -> MuxFn (Pred sym) (SimFrame sym ext (CrucibleLang blocks ret) args)
-mergeCrucibleFrame sym muxFns tgt p x0 y0 =
-  case tgt of
-    BlockTarget _b_id -> do
-      let x = fromJustCallFrame x0
-      let y = fromJustCallFrame y0
-      z <- mergeRegs sym muxFns p (x^.frameRegs) (y^.frameRegs)
-      pure $! MF (x & frameRegs .~ z)
-    ReturnTarget -> do
-      let x = fromNothingCallFrame x0
-      let y = fromNothingCallFrame y0
-      RF <$> muxRegEntry sym muxFns p x y
+-- -- | Get the path conditions from the valueFromValue context
+-- vfvConditions :: ValueFromValue p sym ext r a
+--               -> [Pred sym]
+-- vfvConditions c0 =
+--   case c0 of
+--     VFVCall c _ _ -> pathConditions c
+--     VFVPartial c p _ -> p : vfvConditions c
+--     VFVEnd -> []
 
 ------------------------------------------------------------------------
--- pathConditions
-
--- | Return list of conditions along current execution path.
-pathConditions :: ValueFromFrame p sym ext r a
-               -> [Pred sym]
-pathConditions c0 =
-  case c0 of
-    VFFBranch   c _ _ p _ _ -> p : pathConditions c
-    VFFPartial  c p _ _   -> p : pathConditions c
-    VFFEnd vfv            -> vfvConditions vfv
-
--- | Get the path conditions from the valueFromValue context
-vfvConditions :: ValueFromValue p sym ext r a
-              -> [Pred sym]
-vfvConditions c0 =
-  case c0 of
-    VFVCall c _ _ -> pathConditions c
-    VFVPartial c p _ -> p : vfvConditions c
-    VFVEnd -> []
-
-------------------------------------------------------------------------
--- Resuming frames
-
--- | Resume a paused frame.
-resumeFrame :: PausedPartialFrame p sym ext r f a
-            -> ValueFromFrame p sym ext r f
-            -> ExecCont p sym ext r g b
-resumeFrame pv ctx =
-    withReaderT
-       (stateTree .~ ActiveTree ctx (pv^.pausedValue))
-       (resume pv)
-{-# INLINABLE resumeFrame #-}
-
-
-------------------------------------------------------------------------
--- Return a value
-
-returnAndMerge :: forall p sym ext rtp blocks r args.
-  IsSymInterface sym =>
-  RegEntry sym r ->
-  ExecCont p sym ext rtp (CrucibleLang blocks r) args
-returnAndMerge arg =
-  let cont :: ExecCont p sym ext rtp (CrucibleLang b r) 'Nothing
-      cont = do RF v <- view (stateTree.actFrame.gpValue)
-                returnValue v
-  in
-  withReaderT
-    (stateTree.actFrame.gpValue .~ RF arg)
-    (checkForIntraFrameMerge cont ReturnTarget)
-
--- | Return value from current execution.
---
--- NOTE: All symbolic branching must be resolved before calling `returnValue`.
-returnValue :: IsSymInterface sym
-            => RegEntry sym (FrameRetType f)
-            -> ExecCont p sym ext r f a
-returnValue v =
-  do ActiveTree ctx er <- view stateTree
-     let val = er & partialValue . gpValue .~ v
-     handleSimReturn (returnContext ctx) val
-
-
-handleSimReturn ::
-  IsSymInterface sym =>
-  ValueFromValue p sym ext r ret {- ^ Context to return to. -} ->
-  PartialResult sym ext ret {- ^ Value that is being returned. -} ->
-  ExecCont p sym ext r f a
-handleSimReturn ctx0 return_value = do
-  case ctx0 of
-    VFVCall ctx (MF f) (ReturnToCrucible tpr rest) ->
-      do let v  = return_value^.partialValue.gpValue
-             f' = extendFrame tpr (regValue v) rest f
-         withReaderT
-           (stateTree .~ ActiveTree ctx (return_value & partialValue . gpValue .~ MF f'))
-           continue
-
-    VFVCall ctx _ TailReturnToCrucible ->
-      do let v  = return_value^.partialValue.gpValue
-         withReaderT
-           (stateTree .~ ActiveTree ctx (return_value & partialValue . gpValue .~ RF v))
-           (returnAndMerge v)
-
-    VFVCall ctx (OF f) (ReturnToOverride k) ->
-      do let v = return_value^.partialValue.gpValue
-         withReaderT
-           (stateTree .~ ActiveTree ctx (return_value & partialValue . gpValue .~ OF f))
-           (k v)
-
-    VFVPartial ctx p r ->
-      do sym <- asks stateSymInterface
-         new_ret_val <- liftIO (mergePartialAndAbortedResult sym p return_value r)
-         handleSimReturn ctx new_ret_val
-
-    VFVEnd ->
-      do res <- view stateContext
-         return $! ResultState $ FinishedResult res return_value
-
-
-
-
-
-
-------------------------------------------------------------------------
--- | Checking for intra-frame merge.
-
--- | Return branch target if there is one.
-getIntraFrameBranchTarget ::
-  ValueFromFrame p sym ext root (CrucibleLang b a) ->
-  Maybe (Some (CrucibleBranchTarget b))
-getIntraFrameBranchTarget vff0 =
-  case vff0 of
-  VFFBranch _ _ _ _ _ tgt -> Just (Some tgt)
-  VFFPartial ctx _ _ _ -> getIntraFrameBranchTarget ctx
-  VFFEnd{} -> Nothing
-
-abortPartialResult ::
-  IsSymInterface sym =>
-  SimState p sym ext r f args ->
-  PartialResult sym ext (SimFrame sym ext (CrucibleLang b r') a') ->
-  IO (PartialResult sym ext (SimFrame sym ext (CrucibleLang b r') a'))
-abortPartialResult s pr =
-  let sym                    = stateSymInterface s
-      muxFns                 = stateIntrinsicTypes s
-      abtGp (GlobalPair v g) = GlobalPair <$> popCrucibleFrame sym muxFns v
-                                          <*> globalAbortBranch sym muxFns g
-  in partialValue abtGp pr
-
--- | Branch with a merge point inside this frame.
---
--- If the argument of type @Maybe ('BlockID' b next_args)@ is @Just bid@, and
--- the target of @bid@ is the same as the post-dominator block, then we will
--- check for the opportunity to merge with any pending symbolic branch frames.
--- When all pending merges are complete, run the given computation. If we are
--- not currently about to execute the post-dominator block for a pending merge,
--- then simply run the given computation.
-tryIntraFrameMerge
-  :: forall p sym ext root b r next_args state_args pd_args
-   . IsSymInterface sym
-  => ExecCont p sym ext root (CrucibleLang b r) ('Just next_args)
-     -- ^ What to run for next computation
-  -> CrucibleBranchTarget b pd_args
-     -- ^ The location of the post-dominator block.
-  -> (SimState p sym ext root (CrucibleLang b r) state_args -> SimState p sym ext root (CrucibleLang b r) ('Just next_args))
-    -- ^ Operation to perform on the current tree.
-  -> Maybe (BlockID b next_args)
-    -- ^ 'Just' the frame corresponds to a label, 'Nothing' otherwise
-  -> ExecCont p sym ext root (CrucibleLang b r) state_args
-tryIntraFrameMerge active_cont tgt setter m_bid
-  | Just bid  <- m_bid
-  , Just Refl <- testEquality tgt (BlockTarget bid)
-  = withReaderT setter (checkForIntraFrameMerge active_cont tgt)
-  | otherwise
-  = withReaderT setter active_cont
-
-
--- | Checks for the opportunity to merge within a frame.
---
--- This should be called everytime the current control flow location changes.
-checkForIntraFrameMerge ::
-  ExecCont p sym ext root (CrucibleLang b r) args
-  {- ^ What to run for next computation -} ->
-
-  CrucibleBranchTarget b args
-  {- ^ The location of the current block.-} ->
-
-  ExecCont p sym ext root (CrucibleLang b r) args
-checkForIntraFrameMerge cont tgt =
-  ReaderT $ return . ControlTransferState cont tgt
-
-
--- | Perform a single instance of path merging at a join point.
---   This will resume an alternate branch, if it is pending,
---   or merge result values if a completed branch has alread reached
---   this point. If there are no pending merge points at this location,
---   continue executing by entering the given continuation.
-performIntraFrameMerge ::
-  IsSymInterface sym =>
-  ExecCont p sym ext root (CrucibleLang b r) args
-  {- ^ What to run for next computation -} ->
-
-  CrucibleBranchTarget b args
-  {- ^ The location of the current block.-} ->
-
-  ExecCont p sym ext root (CrucibleLang b r) args
-
-performIntraFrameMerge active_cont tgt = do
-  s <- ask
-  ActiveTree ctx0 er <- view stateTree
-  sym <- asks stateSymInterface
-  case ctx0 of
-    VFFBranch ctx assume_frame loc p other_branch tgt'
-
-      -- Did we get to our merge point (i.e., we are finished with this branch)
-      | Just Refl <- testEquality tgt tgt' ->
-
-        case other_branch of
-
-          -- We still have some more work to do.
-          VFFActivePath toTgt next ->
-            do pathAssumes      <- liftIO $ popAssumptionFrame sym assume_frame
-               new_assume_frame <- liftIO $ pushAssumptionFrame sym
-               pnot             <- liftIO $ notPred sym p
-               liftIO $ addAssumption sym (LabeledPred pnot (ExploringAPath loc toTgt))
-
-               -- The current branch is done
-               let other = VFFCompletePath
-                              pathAssumes
-                              PausedValue { _pausedValue = er
-                                          , resume = active_cont
-                                          }
-               resumeFrame next (VFFBranch ctx new_assume_frame loc pnot other tgt)
-
-          -- We are done with both branches, pop-off back to the outer context.
-          VFFCompletePath otherAssumes other ->
-            do ar <- liftIO $ mergePartialResult s tgt p er (other^.pausedValue)
-
-               -- Merge the assumptions from each branch and add to the
-               -- current assumption frame
-               pathAssumes <- liftIO $ popAssumptionFrame sym assume_frame
-
-               mergedAssumes <- liftIO $ mergeAssumptions sym p pathAssumes otherAssumes
-               liftIO $ addAssumptions sym mergedAssumes
-
-               -- Check for more potential merge targets.
-               withReaderT
-                 (stateTree .~ ActiveTree ctx ar)
-                 (checkForIntraFrameMerge active_cont tgt)
-
-    -- Since the other branch aborted before it got to the merge point,
-    -- we merge-in the partiality on our current path and keep going.
-    VFFPartial ctx p ar needsAborting ->
-      do er'  <- case needsAborting of
-                   NoNeedToAbort    -> return er
-                   NeedsToBeAborted -> liftIO $ abortPartialResult s er
-         er'' <- liftIO $ mergePartialAndAbortedResult sym p er' ar
-         withReaderT
-           (stateTree .~ ActiveTree ctx er'')
-           (checkForIntraFrameMerge active_cont tgt)
-
-    _ -> active_cont
-
-
---------------------------------------------------------------------------------
--- Branching
-
-newtype PausedFrame p sym ext root f args
-      = PausedFrame (PausedValue p sym ext root f args (TopFrame sym ext f args))
-
-pushBranchVal :: (f ~ CrucibleLang b a, IsSymInterface sym)
-              => SimState p sym ext root f ma
-              -> PausedFrame p sym ext root f args
-              -> IO (PausedFrame p sym ext root f args)
-pushBranchVal s (PausedFrame pv) = do
-  let sym = stateSymInterface s
-  let iTypes = stateIntrinsicTypes s
-  let GlobalPair v gs = pv^.pausedValue
-  v' <- pushCrucibleFrame sym iTypes v
-  gs' <- globalPushBranch sym iTypes gs
-  return $! PausedFrame (pv & pausedValue .~ GlobalPair v' gs')
-
-
--- | 'Some' frame, together with its 'BlockID'.
-data SomeLabel p sym ext r b a =
-  forall args.
-  SomeLabel !(PausedFrame p sym ext r (CrucibleLang b a) ('Just args))
-            !(Maybe (BlockID b args))
-
-getTgtLoc ::
-  SimState p sym ext r (CrucibleLang b a) ('Just dc_args) ->
-  BlockID b y ->
-  ProgramLoc
-getTgtLoc s (BlockID i) = blockLoc (blocks Ctx.! i)
-  where
-  blocks = frameBlockMap (s ^. stateCrucibleFrame)
-
--- | Branch with a merge point inside this frame.
-intra_branch ::
-  IsSymInterface sym =>
-  Pred sym
-  {- ^ Branch condition branch -} ->
-
-  SomeLabel p sym ext r b a
-  {- ^ true branch. -} ->
-
-  SomeLabel p sym ext r b a
-  {- ^ false branch. -} ->
-
-  CrucibleBranchTarget b (args :: Maybe (Ctx CrucibleType))
-  {- ^ Merge point, where both branches meet again. -} ->
-
-  ExecCont p sym ext r (CrucibleLang b a) ('Just dc_args)
-
-intra_branch p t_label f_label tgt = do
-  s <- ask
-  ctx <- asContFrame <$> view stateTree
-  sym <- asks stateSymInterface
-  r <- liftIO $ evalBranch sym p
-
-  case r of
-    SymbolicBranch chosen_branch -> do
-      -- Get correct predicate
-      p' <- liftIO $ predEqConst sym p chosen_branch
-
-      (SomeLabel a_state a_id, SomeLabel o_state o_id) <-
-                      return (swap_unless chosen_branch (t_label, f_label))
-
-      loc <- liftIO $ getCurrentProgramLoc sym
-      let a_loc  = getTgtLoc s <$> a_id
-          o_loc  = getTgtLoc s <$> o_id
-
-
-      a_frame <- liftIO $ pushBranchVal s a_state
-      PausedFrame o_frame <- liftIO $ pushBranchVal s o_state
-
-      assume_frame <- liftIO $ pushAssumptionFrame sym
-      liftIO $ addAssumption sym (LabeledPred p' (ExploringAPath loc a_loc))
-
-      -- Create context for paused frame.
-      let o_tree = o_frame & pausedValue %~ TotalRes
-                           & resumed     .~ (tryIntraFrameMerge
-                                                        (resume o_frame)
-                                                           tgt
-                                                           id
-                                                           o_id)
-
-      let todo = VFFActivePath o_loc o_tree
-          ctx' = VFFBranch ctx assume_frame loc p' todo tgt
-
-      -- Start a_state (where branch pred is p')
-      let PausedFrame pf = a_frame
-          setter = stateTree .~ ActiveTree ctx' (TotalRes (pf^.pausedValue))
-      tryIntraFrameMerge (resume pf) tgt setter a_id
-
-    NoBranch chosen_branch ->
-      do SomeLabel a_state a_id <-
-                      return (if chosen_branch then t_label else f_label)
-         let PausedFrame pf = a_state
-             setter = stateTree .~ ActiveTree ctx (TotalRes (pf^.pausedValue))
-
-         tryIntraFrameMerge (resume pf) tgt setter a_id
-
-
-{-# INLINABLE intra_branch #-}
-
-
-
-{- | Merge the assumptions collected from the branches of a conditional.
-The result is a bunch of qualified assumptions: if the branch condition
-is @p@, then the true assumptions become @p => a@, while the false ones
-beome @not p => a@.
--}
-mergeAssumptions ::
-  IsExprBuilder sym =>
-  sym ->
-  Pred sym ->
-  Seq (Assumption sym) ->
-  Seq (Assumption sym) ->
-  IO (Seq (Assumption sym))
-mergeAssumptions sym p thens elses =
-  do pnot <- notPred sym p
-     th' <- (traverse.labeledPred) (impliesPred sym p) thens
-     el' <- (traverse.labeledPred) (impliesPred sym pnot) elses
-     let xs = th' <> el'
-     -- Filter out all the trivally true assumptions
-     return (Seq.filter ((/= Just True) . asConstantPred . view labeledPred) xs)
-
-------------------------------------------------------------------------
--- ValueFromFrame
-
--- | Returns true if tree contains a single non-aborted execution.
-isSingleCont :: ValueFromFrame p sym ext root a -> Bool
-isSingleCont c0 =
-  case c0 of
-    VFFBranch{} -> False
-    VFFPartial c _ _ _ -> isSingleCont c
-    VFFEnd vfv -> isSingleVFV vfv
-
-isSingleVFV :: ValueFromValue p sym ext r a -> Bool
-isSingleVFV c0 = do
-  case c0 of
-    VFVCall c _ _ -> isSingleCont c
-    VFVPartial c _ _ -> isSingleVFV c
-    VFVEnd -> True
-
--- | Attempt to unwind a frame context into a value context.
---   This succeeds only if there are no pending symbolic
---   merges.
-unwindContext :: ValueFromFrame p sym ext root f
-              -> Maybe (ValueFromValue p sym ext root (RegEntry sym (FrameRetType f)))
-unwindContext c0 =
-    case c0 of
-      VFFBranch{} -> Nothing
-      VFFPartial _ _ _ NeedsToBeAborted -> Nothing
-      VFFPartial d p ar NoNeedToAbort ->
-        (\d' -> VFVPartial d' p ar) <$> unwindContext d
-      VFFEnd vfv -> return vfv
-
--- | Get the context for when returning (assumes no
--- intra-procedural merges are possible).
-returnContext :: ValueFromFrame ctx sym ext root f
-              -> ValueFromValue ctx sym ext root (RegEntry sym (FrameRetType f))
-returnContext c0 =
-    case unwindContext c0 of
-      Just vfv -> vfv
-      Nothing ->
-        panic "ExecutionTree.returnContext"
-          [ "Unexpected attempt to exit function before all intra-procedural merges are complete."
-          , "The call stack was:"
-          , show (PP.pretty c0)
-          ]
-
--- | Replace the given frame with a new frame.  Succeeds
---   only if there are no pending symbolic merge points.
-replaceTailFrame :: forall p sym ext a b c args args'
-                  . FrameRetType a ~ FrameRetType c
-                 => ActiveTree p sym ext b a args
-                 -> SimFrame sym ext c args'
-                 -> Maybe (ActiveTree p sym ext b c args')
-replaceTailFrame (ActiveTree c er) f = do
-    vfv <- unwindContext c
-    return $ ActiveTree (VFFEnd vfv) (er & partialValue . gpValue .~ f)
+-- ReturnHandler
+
+-- | This type names a simulator frame, paired with a continuation
+-- to be used when returning from the function.
+
+data ReturnHandler top_return p sym ext r f args new_args where
+  ReturnToOverride ::
+    (ret -> ExecCont p sym ext rtp (OverrideLang args r) 'Nothing)
+      {- Remaining override code to run when the return value becomse available -} ->
+    ReturnHandler ret p sym ext rtp (OverrideLang args r) 'Nothing 'Nothing
+
+  ReturnToCrucible ::
+    TypeRepr ret                       {- Type of the return value -} ->
+    StmtSeq ext blocks r (ctx ::> ret) {- Remaining statements to execute -} ->
+    ReturnHandler (RegEntry sym ret)
+      p sym ext root (CrucibleLang blocks r) ('Just ctx) ('Just (ctx ::> ret))
+
+  TailReturnToCrucible ::
+    ReturnHandler (RegEntry sym r)
+      p sym ext root (CrucibleLang blocks r) ctx 'Nothing
+
+
+returnToOverride ::
+  (ret -> SimState p sym ext rtp (OverrideLang args r) 'Nothing -> IO (ExecState p sym ext rtp))
+    {- ^ Remaining override code to run when the return value becomse available -} ->
+  ReturnHandler ret p sym ext rtp (OverrideLang args r) 'Nothing 'Nothing
+returnToOverride c = ReturnToOverride (ReaderT . c)
+
+returnToCrucible :: forall p sym ext root ret blocks r ctx.
+  TypeRepr ret {- ^ Type of the return value -} ->
+  StmtSeq ext blocks r (ctx ::> ret) {- ^ Remaining statements to execute -} ->
+  ReturnHandler (RegEntry sym ret)
+    p sym ext root (CrucibleLang blocks r) ('Just ctx) ('Just (ctx ::> ret))
+returnToCrucible = ReturnToCrucible
+
+tailReturnToCrucible :: forall p sym ext root blocks ctx r.
+  ReturnHandler (RegEntry sym r)
+    p sym ext root (CrucibleLang blocks r) ctx 'Nothing
+tailReturnToCrucible = TailReturnToCrucible
 
 ------------------------------------------------------------------------
 -- ActiveTree
+
+-- | An active execution tree contains at least one active execution.
+-- The data structure is organized so that the current execution
+-- can be accessed rapidly.
+data ActiveTree p sym ext root (f :: *) args
+   = ActiveTree
+      { _actContext :: !(ValueFromFrame p sym ext root f)
+      , _actResult  :: !(PartialResult sym ext (SimFrame sym ext f args))
+      }
 
 -- | Create a tree with a single top frame.
 singletonTree :: TopFrame sym ext f args
@@ -1246,102 +737,6 @@ activeFrames (ActiveTree ctx ar) =
   SomeFrame (ar^.partialValue^.gpValue) : parentFrames ctx
 
 
-callFn ::
-  ReturnHandler (RegEntry sym (FrameRetType a)) p sym ext r f old_args new_args
-    {- ^ What to do with the result of the function -} ->
-
-  SimFrame sym ext a args
-    {- ^ The code to run -} ->
-
-  ActiveTree p sym ext r f old_args ->
-  ActiveTree p sym ext r a args
-callFn rh f' (ActiveTree ctx er) =
-    ActiveTree (VFFEnd (VFVCall ctx old_frame rh)) er'
-  where
-  old_frame = er ^. partialValue ^. gpValue
-  er'       = er &  partialValue  . gpValue .~ f'
-
-
-------------------------------------------------------------------------
--- Aborting an execution.
-
--- | Abort the current execution, and either return or switch to next
--- execution path and run it.
-abortExec :: IsSymInterface sym
-          => AbortExecReason
-          -> ExecCont p sym ext (r :: *) f args
-abortExec rsn = do
-  ActiveTree ctx ar0 <- view stateTree
-  -- Get aborted result from active result.
-  let ar = case ar0 of
-             TotalRes e -> AbortedExec rsn e
-             PartialRes c ex ar1 -> AbortedBranch c (AbortedExec rsn ex) ar1
-  resumeValueFromFrameAbort ctx ar
-
--- | Resolve the fact that the current branch aborted.
-resumeValueFromFrameAbort ::
-  IsSymInterface sym =>
-  ValueFromFrame p sym ext r f ->
-  AbortedResult sym ext {- ^ The execution that is being aborted. -} ->
-  ExecCont p sym ext r g args
-resumeValueFromFrameAbort ctx0 ar0 = do
-  sym <- asks stateSymInterface
-  case ctx0 of
-
-    -- This is the first abort.
-    VFFBranch ctx assume_frame loc p other_branch tgt ->
-      do pnot <- liftIO $ notPred sym p
-         let nextCtx = VFFPartial ctx pnot ar0 NeedsToBeAborted
-
-         -- Reset the backend path state
-         _assumes <- liftIO $ popAssumptionFrame sym assume_frame
-
-         case other_branch of
-
-           -- We have some more work to do.
-           VFFActivePath toLoc n ->
-             do liftIO $ addAssumption sym (LabeledPred pnot (ExploringAPath loc toLoc))
-                resumeFrame n nextCtx
-
-           -- The other branch had finished successfully;
-           -- Since this one aborted, then the other one is really the only
-           -- viable option we have, and so we commit to it.
-           VFFCompletePath otherAssumes pv ->
-             do let er = pv^.pausedValue
-
-                -- We are committed to the other path,
-                -- assume all of its suspended assumptions
-                liftIO $ addAssumptions sym otherAssumes
-
-                -- check for further merges, then continue onward.
-                withReaderT
-                  (stateTree .~ ActiveTree nextCtx er)
-                  (checkForIntraFrameMerge (resume pv) tgt)
-
-    -- Both branches aborted
-    VFFPartial ctx p ay _ ->
-      resumeValueFromFrameAbort ctx (AbortedBranch p ar0 ay)
-
-    VFFEnd ctx ->
-      resumeValueFromValueAbort ctx ar0
-
--- | Run rest of execution given a value from value context and an aborted
--- result.
-resumeValueFromValueAbort :: IsSymInterface sym
-                          => ValueFromValue p sym ext r ret'
-                          -> AbortedResult sym ext
-                          -> ExecCont p sym ext r f a
-resumeValueFromValueAbort ctx0 ar0 =
-  case ctx0 of
-    VFVCall ctx _ _ -> do
-      -- Pop out of call context.
-      resumeValueFromFrameAbort ctx ar0
-    VFVPartial ctx p ay -> do
-      resumeValueFromValueAbort ctx (AbortedBranch p ar0 ay)
-    VFVEnd ->
-      do res <- view stateContext
-         return $! ResultState $ AbortedResult res ar0
-
 ------------------------------------------------------------------------
 -- extractCurrentPath
 
@@ -1373,59 +768,72 @@ vfvSingleContext ctx0 =
 ------------------------------------------------------------------------
 -- branchConditions
 
--- | Return all branch conditions along path to this node.
-branchConditions :: ActiveTree ctx sym ext ret f args -> [Pred sym]
-branchConditions t =
-  case t^.actResult of
-    TotalRes _ -> vffBranchConditions (t^.actContext)
-    PartialRes p _ _ -> p : vffBranchConditions (t^.actContext)
+-- -- | Return all branch conditions along path to this node.
+-- branchConditions :: ActiveTree ctx sym ext ret f args -> [Pred sym]
+-- branchConditions t =
+--   case t^.actResult of
+--     TotalRes _ -> vffBranchConditions (t^.actContext)
+--     PartialRes p _ _ -> p : vffBranchConditions (t^.actContext)
 
-vffBranchConditions :: ValueFromFrame p sym ext ret f
-                    -> [Pred sym]
-vffBranchConditions ctx0 =
-  case ctx0 of
-    VFFBranch   ctx _ _ p _ _  -> p : vffBranchConditions ctx
-    VFFPartial  ctx p _ _      -> p : vffBranchConditions ctx
-    VFFEnd  ctx -> vfvBranchConditions ctx
+-- vffBranchConditions :: ValueFromFrame p sym ext ret f
+--                     -> [Pred sym]
+-- vffBranchConditions ctx0 =
+--   case ctx0 of
+--     VFFBranch   ctx _ _ p _ _  -> p : vffBranchConditions ctx
+--     VFFPartial  ctx p _ _      -> p : vffBranchConditions ctx
+--     VFFEnd  ctx -> vfvBranchConditions ctx
 
-vfvBranchConditions :: ValueFromValue p sym ext root top_ret
-                    -> [Pred sym]
-vfvBranchConditions ctx0 =
-  case ctx0 of
-    VFVCall     ctx _ _      -> vffBranchConditions ctx
-    VFVPartial  ctx p _      -> p : vfvBranchConditions ctx
-    VFVEnd                   -> []
+-- vfvBranchConditions :: ValueFromValue p sym ext root top_ret
+--                     -> [Pred sym]
+-- vfvBranchConditions ctx0 =
+--   case ctx0 of
+--     VFVCall     ctx _ _      -> vffBranchConditions ctx
+--     VFVPartial  ctx p _      -> p : vfvBranchConditions ctx
+--     VFVEnd                   -> []
+
 
 
 ------------------------------------------------------------------------
--- FrameRetType
+-- SimContext
 
-type family FrameRetType (f :: *) :: CrucibleType where
-  FrameRetType (CrucibleLang b r) = r
-  FrameRetType (OverrideLang b r) = r
+-- | A definition of a function's semantics, given as a Haskell action.
+data Override p sym ext (args :: Ctx CrucibleType) ret
+   = Override { overrideName    :: FunctionName
+              , overrideHandler :: forall r. ExecCont p sym ext r (OverrideLang args ret) 'Nothing
+              }
 
-type CrucibleState p sym ext rtp blocks ret args
-   = SimState p sym ext rtp (CrucibleLang blocks ret) ('Just args)
+-- | State used to indicate what to do when function is called.
+data FnState p sym ext (args :: Ctx CrucibleType) (ret :: CrucibleType)
+   = UseOverride !(Override p sym ext args ret)
+   | forall blocks . UseCFG !(CFG ext blocks args ret) !(CFGPostdom blocks)
 
 -- | A map from function handles to their semantics.
 type FunctionBindings p sym ext = FnHandleMap (FnState p sym ext)
 
+-- | The type of functions that interpret extension statements.  These
+--   have access to the main simulator state, and can make fairly arbitrary
+--   changes to it.
 type EvalStmtFunc p sym ext =
   forall rtp blocks r ctx tp'.
     StmtExtension ext (RegEntry sym) tp' ->
     CrucibleState p sym ext rtp blocks r ctx ->
     IO (RegValue sym tp', CrucibleState p sym ext rtp blocks r ctx)
 
+-- | In order to start executing a simulator, one must provide an implementation
+--   of the extension syntax.  This includes an evaluator for the added
+--   expression forms, and an evaluator for the added statement forms.
 data ExtensionImpl p sym ext
   = ExtensionImpl
     { extensionEval :: IsSymInterface sym => sym -> IntrinsicTypes sym -> (Int -> String -> IO ()) -> EvalAppFunc sym (ExprExtension ext)
     , extensionExec :: EvalStmtFunc p sym ext
     }
 
-------------------------------------------------------------------------
--- SimContext
+type IsSymInterfaceProof sym a = (IsSymInterface sym => a) -> a
 
--- | Global context for state.
+-- | Top-level state record for the simulator.  The state contained in this record
+--   remains persistent across all symbolic simulator actions.  In particular, it
+--   is not rolled back when the simulator returns previous program points to
+--   explore additional paths, etc.
 data SimContext (personality :: *) (sym :: *) (ext :: *)
    = SimContext { _ctxSymInterface       :: !sym
                   -- | Class dictionary for @'IsSymInterface' sym@
@@ -1440,31 +848,17 @@ data SimContext (personality :: *) (sym :: *) (ext :: *)
                 , _cruciblePersonality   :: !personality
                 }
 
--- | A definition of function semantics.
-data Override p sym ext (args :: Ctx CrucibleType) ret
-   = Override { overrideName :: FunctionName
-              , overrideHandler
-                  :: forall r. ExecCont p sym ext r (OverrideLang args ret) 'Nothing
-              }
-
--- | State used to indicate what to do when function is called.
-data FnState p sym ext (args :: Ctx CrucibleType) (ret :: CrucibleType)
-   = UseOverride !(Override p sym ext args ret)
-   | forall blocks . UseCFG !(CFG ext blocks args ret) !(CFGPostdom blocks)
-
-------------------------------------------------------------------------
--- SimContext utilities
-
 -- | Create a new 'SimContext' with the given bindings.
-initSimContext :: IsSymInterface sym
-               => sym
-               -> IntrinsicTypes sym
-               -> HandleAllocator RealWorld
-               -> Handle -- ^ Handle to write output to
-               -> FunctionBindings personality sym ext
-               -> ExtensionImpl personality sym ext
-               -> personality
-               -> SimContext personality sym ext
+initSimContext ::
+  IsSymInterface sym =>
+  sym ->
+  IntrinsicTypes sym ->
+  HandleAllocator RealWorld ->
+  Handle {- ^ Handle to write output to -} ->
+  FunctionBindings personality sym ext ->
+  ExtensionImpl personality sym ext ->
+  personality ->
+  SimContext personality sym ext
 initSimContext sym muxFns halloc h bindings extImpl personality =
   SimContext { _ctxSymInterface     = sym
              , ctxSolverProof       = \a -> a
@@ -1490,32 +884,59 @@ cruciblePersonality = lens _cruciblePersonality (\s v -> s{ _cruciblePersonality
 ------------------------------------------------------------------------
 -- SimState
 
-type IsSymInterfaceProof sym a = (IsSymInterface sym => a) -> a
+
+-- | An abort handler indicates to the simulator what actions to take
+--   when an abort occurs.  Usually, one should simply use the
+--   'defaultAbortHandler' from "Lang.Crucible.Simulator", which
+--   unwinds the tree context to the nearest branch point and
+--   correctly resumes simulation.  However, for some use cases, it
+--   may be desirable to take additional or alternate actions on abort
+--   events; in which case, the libary user may replace the default
+--   abort handler with their own.
+newtype AbortHandler p sym ext rtp
+      = AH { runAH :: forall (l :: *) args.
+                 AbortExecReason ->
+                 ExecCont p sym ext rtp l args
+           }
 
 -- | A SimState contains the execution context, an error handler, and
--- the current execution tree.
+--   the current execution tree.
 data SimState p sym ext rtp f (args :: Maybe (Ctx.Ctx CrucibleType))
    = SimState { _stateContext      :: !(SimContext p sym ext)
-              , _errorHandler      :: !(ErrorHandler p sym ext rtp)
+              , _abortHandler      :: !(AbortHandler p sym ext rtp)
               , _stateTree         :: !(ActiveTree p sym ext rtp f args)
               }
 
-newtype ErrorHandler p sym ext rtp
-      = EH { runEH :: forall (l :: *) args
-                    . AbortExecReason
-                   -> ExecCont p sym ext rtp l args
-           }
+-- | A simulator state that is currently executing Crucible instructions.
+type CrucibleState p sym ext rtp blocks ret args
+   = SimState p sym ext rtp (CrucibleLang blocks ret) ('Just args)
 
-------------------------------------------------------------------------
--- SimState lens
+
+-- | Create an initial 'SimState'.
+initSimState :: SimContext p sym ext
+             -> SymGlobalState sym
+             -- ^ state of global variables
+             -> AbortHandler p sym ext (RegEntry sym ret)
+             -> SimState p sym ext (RegEntry sym ret) (OverrideLang EmptyCtx ret) 'Nothing
+initSimState ctx globals ah =
+  let startFrame = OverrideFrame { override = startFunctionName
+                                 , overrideRegMap = emptyRegMap
+                                 }
+      startGP = GlobalPair (OF startFrame) globals
+   in SimState
+      { _stateContext = ctx
+      , _abortHandler = ah
+      , _stateTree    = singletonTree startGP
+      }
 
 -- | View a sim context.
 stateContext :: Simple Lens (SimState p sym ext r f a) (SimContext p sym ext)
 stateContext = lens _stateContext (\s v -> s { _stateContext = v })
 {-# INLINE stateContext #-}
 
-errorHandler :: Simple Lens (SimState p sym ext r f a) (ErrorHandler p sym ext r)
-errorHandler = lens _errorHandler (\s v -> s { _errorHandler = v })
+-- | Access the current abort handler of a state.
+abortHandler :: Simple Lens (SimState p sym ext r f a) (AbortHandler p sym ext r)
+abortHandler = lens _abortHandler (\s v -> s { _abortHandler = v })
 
 -- | Access the active tree associated with a state.
 stateTree :: Lens (SimState p sym ext rtp f a)
@@ -1524,6 +945,21 @@ stateTree :: Lens (SimState p sym ext rtp f a)
                   (ActiveTree p sym ext rtp g b)
 stateTree = lens _stateTree (\s v -> s { _stateTree = v })
 {-# INLINE stateTree #-}
+
+-- | A lens to access the Crucible call frame inside a 'SimState'
+stateCrucibleFrame ::
+  Lens (SimState p sym ext rtp (CrucibleLang blocks r) ('Just a))
+       (SimState p sym ext rtp (CrucibleLang blocks r) ('Just a'))
+       (CallFrame sym ext blocks r a)
+       (CallFrame sym ext blocks r a')
+stateCrucibleFrame = stateTree . actFrame . crucibleTopFrame
+{-# INLINE stateCrucibleFrame #-}
+
+stateOverrideFrame :: Lens (SimState p sym ext q (OverrideLang a r) 'Nothing)
+                           (SimState p sym ext q (OverrideLang a r) 'Nothing)
+                           (OverrideFrame sym r a)
+                           (OverrideFrame sym r a)
+stateOverrideFrame = stateTree . actFrame . gpValue . overrideSimFrame
 
 stateSymInterface :: SimState p sym ext r f a -> sym
 stateSymInterface s = s ^. stateContext . ctxSymInterface
@@ -1537,30 +973,30 @@ stateIntrinsicTypes s = ctxIntrinsicTypes (s^.stateContext)
 stateGetConfiguration :: SimState p sym ext r f args -> Config
 stateGetConfiguration s = stateSolverProof s (getConfiguration (stateSymInterface s))
 
-------------------------------------------------------------------------
--- HasSimState instance
 
--- | View a sim context.
-stateOverrideFrame :: Lens (SimState p sym ext q (OverrideLang a r) 'Nothing)
-                           (SimState p sym ext q (OverrideLang a r) 'Nothing)
-                           (OverrideFrame sym r a)
-                           (OverrideFrame sym r a)
-stateOverrideFrame = stateTree . actFrame . gpValue . overrideSimFrame
+{-
+????
 
-------------------------------------------------------------------------
--- Running error handler
-
--- | Start running the error handler.
-runErrorHandler ::
-  AbortExecReason ->
-  ExecCont p sym ext rtp f args
-runErrorHandler err = ReaderT $ \s -> return (AbortState err s)
-
-runGenericErrorHandler ::
-  String ->
-  ExecCont p sym ext rtp f args
-runGenericErrorHandler msg =
-  do ctx <- view stateContext
-     let sym = ctx^.ctxSymInterface
-     loc <- ctxSolverProof ctx (liftIO (getCurrentProgramLoc sym))
-     runErrorHandler (ManualAbort loc msg)
+checkStateConsistency :: CrucibleState p sym rtp blocks r a
+                      -> BlockID blocks args
+                         -- ^ Current block of top of stack frame.
+                      -> IO ()
+checkStateConsistency s (BlockID block_id) = do
+  let f = s^.stateCrucibleFrame
+  case getIntraFrameBranchTarget (s^.stateTree^.actContext) of
+    Nothing -> return ()
+    Just (Some tgt) -> do
+      let Const _pd = framePostdomMap f Ctx.! block_id
+      case branchTarget tgt of
+        ReturnTarget -> return ()
+          -- FIXME? I'm not sure ignoring this situation is correct...
+          -- unless (null pd) $ do
+          --   fail $ "The crucible tree reached an illegal state.\n"
+          --       ++ "Branch target: Function return\n"
+          --       ++ "Postdoms:      " ++ show pd
+        BlockTarget _tgt' -> return ()
+          -- when (Some tgt' `notElem` pd) $ do
+          --   fail $ "The crucible tree reached an illegal state.\n"
+          --       ++ "Branch target: " ++ show tgt' ++ "\n"
+          --       ++ "Postdoms:      " ++ show pd
+-}
